requirements.txt
annotated-types
anthropic
anyio
asyncpg
attrs
alembic
certifi
cffi
charset-normalizer
click
colorama
distro
fastapi
filelock
fsspec
greenlet
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
idna
iniconfig
jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain-core
langchain-text-splitters
langsmith
markdown-it-py
markupsafe
mcp
mdurl
mpmath
networkx
nltk
numpy
orjson
packaging
pymilvus
pgvector
pillow
pluggy
psycopg2-binary
pycparser
pydantic
pydantic-core
pydantic-settings
pygit2
pygments
pytest
python-dotenv
python-multipart
pywin32
pyyaml
referencing
regex
requests
requests-toolbelt
rich
rpds-py
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
sniffio
sqlalchemy
sse-starlette
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
typer
typing-extensions
typing-inspection
urllib3
uvicorn
watchfiles
websockets
zstandard
